Use pip instead of conda in readthedocs
Prior to this commit, readthedocs uses conda to setup the environment
and to install narupa. This leads to 2 issues:

* building a conda environment can use more memory than what is allowed
on readthedocs;
* issues with uploading the conda packages may lead to synchronisation
mismatch between the code in the repo and the doc.

Here, we switch to pip to build the environment and revert to compile.sh
to install narupa.

By using pip to build the environment, we encounter issues when
installing OpenMM. OpenMM is not installable from pip, also the name
openmm is used for a different tool on pypi so `pip install openmm`
installs the wrong library. Though, OpenMM needs to be importable for
the documentation to be built. This is fixed by introducing a mock
library for OpenMM.

The openmm mock library silences the warning in the compile script, but
also avoid import errors when building the doc. The openmm mock library
lets you import anything from the fake simtk package; any name starting
in lower case (e.g simtk.anything.name) is imported as a package, any
name starting in upper case (e.g. simtk.something.boo.Name) is imported
as a dummy class. This openmm mock relies on the "pretends" library that
implements the mocking logic for the sake of this commit.

The mpi4py mock library assures that mpi4py appears as installed to
silence the compile script on linux. Note that I did not check if MPI
was available in the readthedocs environemnt, so it may be possible to
just install the actual mpi4py library.

The use of the compile.sh script to install narupa did break at some
point, leading to the use of the conda packages. Fixing that breakage
required to:

* assure that the compile script was run in no-edit mode
* fix the system search path so that the compile script uses the
expected version of python