Highlights taken out of homepage and put onto own page
Added a high-level description of NanoVer to the home page and moved the highlights sections that were on the home page to a separate page called "Highlights".  The homepage feels a bit cluttered...I'm happy to hear suggestions!

For the highlights, I've rearranged the info and added a gif and video. Note that the video doesn't play if you just open the built html, you have to run "python -m http.server" and then open http://localhost:8000 in your browser. (It should play fine on the online version).