Update installation.rst
Added a note about the fact that these instructions might not cover all possible VR configurations and point them to the online documentation for their specific headset if they have problems setting up their VR kit.

Add an installation option (that I recently learned about and that they are using in Bristol) about running the apk on the headset but through the meta developer hub, which means they can connect to a server running on their computer without communication over the network (ideal for eduroam).