Add info about running servers (#80)
* Add contents and update intro

Added subheadings in preparation for adding a section on running a NanoVer server. Also reworded the into to the tutorials part to make it fit better with the new heading.

* Added basics about servers and the rust server

Added a section in Tutorials/Fundamentals with information about NanoVer servers. Added a section about protocol (currently empty) and the rust server (with basic info).

* Add more info

* Added info about "getting started" notebook, which is referenced in the new section about running a server
* Added information about the Python server
* Updated the Rust server section

* Update fundamentals.rst

Update wording & reformat.

* Refactor to generalise instructions for Rust server across operating systems

* Minor edits

* Delete duplicates and edit code snippets

---------

Co-authored-by: Harry Stroud <harry.stroud@usc.es>