Update flow diagram
The flow diagram stated that if someone is using "other" as their type of VR headset, then they must use SteamVR. This is not correct, they can actually use SteamVR (tethered) or SteamLink (wireless). This has been corrected in the 2nd blue box from the top in the flow diagram (both in the powerpoint source file and the png).