Move changing interaction force scaling on page and fix typo
The instructions on how to change the interaction force scaling were previously nested underneath the "right hand menus" because you use the right controller to do this. However, it is more accurate to nest it under "in-simulation controls" as it is not a 'menu' as such, and can be done anytime the user doesn't have any menus open.