Add page about in-VR menus in NanoVer iMD (#127)
* Add videos about connecting to a server

Added a note about how to 'click' in VR with a controller. Also added explanations and three videos for connecting to a server via autoconnect, manual connect, and discover services. The videos were added as mp4s rather than gifs (gifs were too low quality), which required the `sphinxcontrib.video` package. This package has been added to the conf file and the requirements file (you will need to reinstall the required packages).

* Add contents at the top of page

* Change contents to local on NanoVer iMD tutorial page

The contents at the top of this page was not set to :local:, meaning that the page title was in the content which is unnecessary.

* Add sections for the right hand and left hand controls, and edit first section

Added a section for the right hand controls: changing force scaling & interaction type.
Added a section for the left hand controls: simulation commands, changing simulation, and customizing your avatar.

Added videos for all of these things, plus explanatory text.

* Fix heading underline

* Fix heading underline

* Fix dropdown menu errors

* Changes during review process

* Update menu headings

* nest pages & rename page to be consistent with it's title

The three headings of the page have now been nested into their own pages.

* Add a local contents for the page with multiple headings

* Rearrange and update wording about interacting with the VR menu buttons

* Hide video tutorials in drop downs for Main Menu page

* Move video tutorials into dropdowns

* Move changing interaction force scaling on page and fix typo

The instructions on how to change the interaction force scaling were previously nested underneath the "right hand menus" because you use the right controller to do this. However, it is more accurate to nest it under "in-simulation controls" as it is not a 'menu' as such, and can be done anytime the user doesn't have any menus open.

* Add directives to open the dropdowns

* Update wording

* checkpoint current state

It's a bit of a mess at the moment. I'm going to save the current state and then totally mess around with the format.

* Put everything back onto the NanoVer iMD Tutorials page

I have moved everything back onto the main NanoVer iMD Tutorial page. I have tried to condense the information & make less overwhelming using: dropdown menus, forced line breaks, and horizontal lines.

I have updated the headings to make them clearer & more concise.

I have added a section called "A typical iMD-VR workflow" with step-by-step instructions on a basic use of NanoVer iMD, with links to relevant sections for more detailed instructions.

I removed old pages as these are now contained in the vrclient.rst page.

* Fix incorrect dropdown for "user" in the desktop menu section

I had accidentally copied the wrong info into this section.

* Add screenshots of menus, remove lines between subheadings & update wording

* Removed the horizontal line spacers between subheadings. We now only use these to separate headings.
* For the in-VR menus section, go from these subheadings: [in-simulation controls, right hand menus & left hand menus] to these: [in-simulation controls, right handheld menu, left handheld menu & fullscreen menu]. This removes a sublayer of sections and is less confusing to the reader. Note that the fullscreen menu is accessed via the left handheld menu, but this is explained within it's own section and this is why the fullscreen section comes after the left handheld menu.
* Listed the above subheadings underneath the "VR controls & menus" heading, which should make navigation easier.
* Added screenshots for the VR menus to be visible without needed to open the dropdown menus. I think this splits the page better into sections and makes it clearer to the reader.
* Changed "full screen" to "fullscreen".
* Made sure that the formatting was applied consistently to the "Choosing your iMD-VR setup" section at the bottom of the page.
* Updated the wording of the "Choosing your iMD-VR setup" and added a step-by-step list of instructions for using this section.
* Added a sentence at the top of the page navigating readers to the two sections of the page.
* Generally updated wording for clearer reading.

* Minor wording changes

Updated the wording on various parts of the page for clarity.

* Update "NanoVer iMD" to "Nanover iMD-VR"

* Update URLs to nanover-imd-vr repo

* Fix formatting issue

* Update subheading level

* Update forgotten nanover-imd reference

---------

Co-authored-by: Harry Stroud <harry.stroud@usc.es>