Update docs release date (#152)
* Update docs release date

Previously, the old release date was set on when the nanover-imd-vr repo was built. I have changed this to be when the docs is built.

I had a look at other software's documentation and they didn't have the release date in the title. So I removed the release date from the html title and moved it to the footer of the page. The title is now just "[page title] - NanoVer documentation". This formatting is the same that ASE uses.

I have added a todo to say that we should add the version number of nanover-server-py to the html title.

* Add pytz package to requirements