Add option to use mobile hotspot instead of Eduroam (#176)
* Add mobile hotspot as option to flow diagram

Modified the flow diagram for "choosing your iMD-VR setup" to specify that private networks include both private Wi-Fi and mobile hotspots. This change is reflected in the png image and the powerpoint source file. Also added headings to the powerpoint to make it clear what is on each page.

* Add URL to NanoVer-iMD-VR tutorial page

Add hyperlink to nanover-imd-vr github page.

* Clarify use of mobile hotspot for wireless setup

Add clarification to the "requirements for a wireless setup" that you may use a mobile hotspot.