Move info about processing NanoVer recordings to Tutorials (#188)
* Move info about recording in NanoVer and processing recordings to Tutorials

Move the information about how to record with NanoVer and processing NanoVer recordings from Concepts to Tutorials.

Added a section at the bottom of the page (Tutorials), ready to put information about how to use NanoVer utilities. For example, using nanover-state-dump, stripping fields from frames, and splitting recordings.

* Remove unused heading

Remove unused heading, this can be added in when that section is written.

* Update wording and add clearer introductions

* Add reference to concepts page on the tutorials page

* Update basics.rst

Update wording to make sense on the new page.